requirements.txt
celery==5.3.4
django==5.1.5
django_bootstrap3==24.3
djangorestframework==3.15.2
drf_yasg==1.21.8
google-api-python-client==2.160.0
oauth2client==4.1.3
psycopg2_binary
python_telegram_bot==21.10
python-keycloak==5.1.1
redis==5.0.0
requests==2.32.3
sentry-sdk[django,celery]==2.20.0
uwsgi==2.0.28
